requirements.txt
# ============================================================
# FuelGuard — Transport Fuel Theft Prevention Platform
# requirements.txt  |  Python 3.10+
# Install: pip install -r requirements.txt
# ============================================================

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# MongoDB Async Driver
motor==3.7.0
pymongo==4.11.1

# Data Validation
pydantic==2.10.6
pydantic-settings==2.8.0

# OCR and Image Processing
opencv-python==4.11.0.86
pytesseract==0.3.13
Pillow==11.1.0
numpy==2.2.3

# Fleet Manager Dashboard
streamlit==1.43.2
plotly==6.0.0
pandas==2.2.3

# Auth and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Notifications
twilio==9.4.5

# Utilities
python-dotenv==1.0.1
httpx==0.28.1

# LLM Vision (Gemini)
google-generativeai

# Dev and Testing
pytest==8.3.5
pytest-asyncio==0.25.3
